+++ packages/bridge-nest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.7.0
+
 ## 0.2.4
 
 ### Patch Changes