+++ packages/health-check-nest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.7.0
+  - @volontariapp/bridge-nest@0.2.5
+
 ## 0.1.8
 
 ### Patch Changes